pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-graph-rag"
version = "0.2.0"
description = "Medical Graph RAG with hierarchical knowledge graph using Ollama LLMs"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Graph RAG Builder" }]
dependencies = [
    "neo4j>=5.18,<6.0",
    "httpx>=0.27",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
    "sentence-transformers>=2.2",
    "numpy>=1.26",
    "rich>=13.7",
    "typer>=0.9",
    "streamlit>=1.32",
    "streamlit-agraph>=0.0.45",
    "pandas>=2.2",
    "pypdf>=4.0",
    "python-docx>=1.1",
    "beautifulsoup4>=4.12",
    "lxml>=5.1",
    "openpyxl>=3.1",
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = ["pytest>=8.1", "ruff>=0.4", "mypy>=1.10"]

[project.scripts]
graph-rag = "graph_rag.cli:app"

[tool.setuptools.packages.find]
include = ["graph_rag*"]
